Refactor: check_image.dart.dart 이름 user_check.dart 로 변경
